필요없는 스크린삭제
